This implements a very limited system for removing blade caches.
All it does is look for a directory in your theme folder, and then
delete all files in that folder. You can tell it which folder to look
by passing a `--directory` argument; otherwise it looks in
`cache/blade`.

Ideally, this should behave more like https://github.com/alwaysblank/blade-generate,
but I haven't been able to figure out how to load in the Blade interface
to use for determining which cached files correspond to which Blades.